requirements.txt
# Core web framework
flask>=2.3.0
flask-cors>=4.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-cohere>=0.1.0

# AI/ML APIs and models
cohere>=4.0.0
sentence-transformers>=2.2.0

# Vector database
chromadb>=0.4.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Testing framework
pytest>=7.0.0

# HTTP requests
requests>=2.28.0

# Environment and configuration
python-dotenv>=1.0.0

# TensorFlow/Keras compatibility (for Keras 3)
tf-keras>=2.15.0

# Additional utilities
pydantic>=2.0.0
typing-extensions>=4.0.0